Organised into seperate files and added config.h
