+++ src/config.h
@@ -1,0 +1,1 @@
+#define WRITE_TO_FILE
