+++ src/file.c
@@ -1,0 +1,13 @@
+#include <assert.h>
+
+// NOT IMPLEMENTED
+int Savetofile(char* pass) {
+	assert(0 && "Not Implemented");
+	return 0;
+}
+
+// NOT IMPLEMENTED
+int CPtoClipboard(char* pass) {
+	assert(0 && "Not Implemented");
+	return 0;
+}
