+++ src/main.h
@@ -1,0 +1,3 @@
+#define MAX_CHAR_TYPE_COMBINATIONS 4
+
+void ShowArguments();
